[tests/*.c] Support the case where mpfr_exp_t is not a long int (but with such a future extension, one should probably provide a MPFR_PRIexp macro, similar to what <inttypes.h> does).
[From SVN r7226 (trunk)]

+++ tests/tcan_round.c
@@ -50,7 +50,7 @@ check_round_p (void)
         {
           printf ("mpfr_round_p(%d) != mpfr_can_round(%d)!\n"
                   "bn = %ld, err0 = %ld, prec = %lu\nbp = ",
-                  r1, r2, n, err, (unsigned long) p);
+                  r1, r2, n, (long) err, (unsigned long) p);
           gmp_printf ("%NX\n", buf, n);
           exit (1);
         }
